Run commands on the server
