+++ src/cli/server.rs
@@ -1,27 +1,37 @@
+extern crate clap;
+extern crate yaml_rust;
+
 use {
-    cli::logger,
+    cli::{cmd, logger},
     std::{
         io::{Read, Write},
         net::{TcpListener, TcpStream},
-        process, thread,
+        process,
+        sync::Arc,
+        thread,
     },
 };
+
+use self::{clap::App, yaml_rust::yaml::Yaml};
 
 const SERVER_ADDRESS: &str = "127.0.0.1:4774";
 const SIG_KILL: &str = "kill\0";
 const SIG_CHALLENGE: &str = "lost\0";
 const SIG_RESPONSE: &str = "found\0";
 
-pub fn start() {
+pub fn start(clap_config: &'static Yaml) {
     logger::info(&format!("Starting padd server on {}", SERVER_ADDRESS));
+
+    let clap_config_arc = Arc::new(clap_config);
 
     match TcpListener::bind(SERVER_ADDRESS) {
         Ok(listener) => {
             for stream in listener.incoming() {
                 match stream {
                     Ok(stream) => {
+                        let config = clap_config_arc.clone();
                         thread::spawn(move || {
-                            handle_stream(stream);
+                            handle_stream(stream, config);
                         });
                     }
                     Err(err) => {
@@ -34,7 +44,7 @@ pub fn start() {
     };
 }
 
-fn handle_stream(stream: TcpStream) {
+fn handle_stream(stream: TcpStream, clap_config: Arc<&Yaml>) {
     match stream.try_clone() {
         Ok(mut stream_writer) => {
             let mut buf: Vec<u8> = Vec::new();
@@ -62,7 +72,7 @@ fn handle_stream(stream: TcpStream) {
                             logger::err(&format!("Failed to write challenge response: {}", err));
                         }
                     }
-                    &_ => execute_command(string),
+                    &_ => execute_command(string, clap_config),
                 },
                 Err(err) => logger::err(&format!(
                     "Failed to convert stream bytes to string: {}",
@@ -105,6 +115,7 @@ pub fn running() -> bool {
 }
 
 pub fn send_command(mut command: String) {
+    // Add terminal character
     command.push('\0');
 
     match TcpStream::connect(SERVER_ADDRESS) {
@@ -117,6 +128,17 @@ pub fn send_command(mut command: String) {
     }
 }
 
-fn execute_command(command: String) {
+fn execute_command(mut command: String, clap_config: Arc<&Yaml>) {
     logger::info(&format!("Executing command: {}", &command));
+
+    // Remove terminal character
+    command.pop();
+
+    let args = command.split_whitespace();
+    let matches = App::from_yaml(&clap_config).get_matches_from(args);
+
+    if let Some(matches) = matches.subcommand_matches("fmt") {
+        println!("Formatting!");
+        cmd::fmt(&matches);
+    }
 }
